+++ src/components/common/fields/InputField.vue
@@ -37,6 +37,10 @@
       {{ label }}
     </span>
 
+    <div class="input-field__hints">
+      <slot name="hint" />
+    </div>
+
     <transition name="input-field__err-transition">
       <p class="input-field__err-mes" v-if="errorMessage">
         {{ errorMessage }}